Rename version.py to _version.py
