+++ noxfile.py
@@ -209,7 +209,7 @@ def check_style(session):
     session.run("black", "--check", ".")
 
 
-@nox.session(reuse_venv=True)
+@nox.session()
 def docs(session):
     try:
         session.run("dot", "-V", silent=True, external=True)
