+++ pycyphal/__init__.py
@@ -34,7 +34,7 @@ the library log level:
 import os as _os
 
 
-from .version import __version__ as __version__
+from ._version import __version__ as __version__
 
 __version_info__ = tuple(map(int, __version__.split(".")[:3]))
 __author__ = "OpenCyphal"
