+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name            = pycyphal
-version         = attr: pycyphal.version.__version__
+version         = attr: pycyphal._version.__version__
 author          = OpenCyphal
 author_email    = consortium@opencyphal.org
 url             = https://opencyphal.org
